dev_requirements.txt
pytest
pytest-cov
pytest_httpserver
responses
freezegun
vcrpy
flake8
black
flask
